Fix auto-submit + change
