+++ src/extensions/views/InstallCustomExtension/components/InstallExtensionManifestData/index.ts
@@ -1,0 +1,1 @@
+export * from "./InstallExtensionManifestData";
