+++ src/extensions/views/InstallCustomExtension/components/InstallSectionData/index.ts
@@ -1,0 +1,1 @@
+export * from "./InstallSectionData";
